construction_report_bot/requirements.txt
aiogram>=3.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
aiosqlite>=0.19.0
flake8>=7.0.0
flake8-docstrings>=1.7.0
flake8-quotes>=3.4.0
psycopg2-binary>=2.9.9
asyncpg>=0.29.0 